pending issues
At the end of the animation it resets to the Wyyard station location at a zoomed in position. It should end of the la-renae... location i.e. the last poistion not reset to the start

the animation does not play between locations see below

this is the sequence explaining what happens

Press play

Wynyard station summary from json displays

Map does not centre but icon group correctly animates along 1st segment from wynyard to the glenmore

 glenmore summary from json displays

...from here on after a pause the summary property for each location displays but there is no display of the guest icons moving along the path

the process ends at Wynyard station as per the issue reported above