new additions 1. add a start dlg box on page load, content coming from a property in the config.json which allows me to present an introduction and explain the controls 2. add an end dlg box 3. whether the start and end dlgs display should depend on a property "show: true" to enable me to display this display 4. the json content should support hyperlinks to open external web pages
E.g. to start

Heading "Welcome to Deb's 60th birthday"
Content " This is a short overview of the route and locations for the day. After our brief presentation you can click on the location markers to see details of the venues with highlighted menu items prices and more!"

this dlg should be dismissable

e.g. to end

Heading "I hope you enjoyed the tour"
Content "Feel free to move about the map and explore!"